feat: whatsmeow ack event support (#362)
* Add webhook support for WhatsApp message acknowledgement events


* Add context parameter to forwardReceiptToWebhook for future-proofing


* refactor(webhook): restructure event handling and add receipt event support

- Split webhook.go into event_message.go, event_receipt.go, and event_delete.go
- Enhanced receipt event handling with detailed payload descriptions
- Updated webhook payload documentation with receipt event examples
- Improved ephemeral message handling in send operations
- Standardized JSON formatting in documentation

* Enhance ephemeral message handling in SQLite repository

- Added logic to preserve existing ephemeral expiration when creating or updating chats.
- Improved error handling when retrieving existing chat data.
- Updated send operation to set default ephemeral expiration if not provided in the request.

---------
