+++ src/infrastructure/whatsapp/webhook.go
@@ -6,210 +6,15 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"regexp"
-	"strings"
 	"time"
 
-	"go.mau.fi/whatsmeow/types"
-
 	"github.com/aldinokemal/go-whatsapp-web-multidevice/config"
-	domainChatStorage "github.com/aldinokemal/go-whatsapp-web-multidevice/domains/chatstorage"
 	pkgError "github.com/aldinokemal/go-whatsapp-web-multidevice/pkg/error"
 	"github.com/aldinokemal/go-whatsapp-web-multidevice/pkg/utils"
 	"github.com/sirupsen/logrus"
-	"go.mau.fi/whatsmeow/types/events"
 )
 
-// forwardToWebhook is a helper function to forward event to webhook url
-func forwardToWebhook(ctx context.Context, evt *events.Message) error {
-	logrus.Info("Forwarding event to webhook:", config.WhatsappWebhook)
-	payload, err := createPayload(ctx, evt)
-	if err != nil {
-		return err
-	}
-
-	for _, url := range config.WhatsappWebhook {
-		if err = submitWebhook(payload, url); err != nil {
-			return err
-		}
-	}
-
-	logrus.Info("Event forwarded to webhook")
-	return nil
-}
-
-func createPayload(ctx context.Context, evt *events.Message) (map[string]any, error) {
-	message := utils.BuildEventMessage(evt)
-	waReaction := utils.BuildEventReaction(evt)
-	forwarded := utils.BuildForwarded(evt)
-
-	body := make(map[string]any)
-
-	body["sender_id"] = evt.Info.Sender.User
-	body["chat_id"] = evt.Info.Chat.User
-
-	if from := evt.Info.SourceString(); from != "" {
-		body["from"] = from
-
-		from_user, from_group := from, ""
-		if strings.Contains(from, " in ") {
-			from_user = strings.Split(from, " in ")[0]
-			from_group = strings.Split(from, " in ")[1]
-		}
-
-		if strings.HasSuffix(from_user, "@lid") {
-			body["from_lid"] = from_user
-			lid, err := types.ParseJID(from_user)
-			if err != nil {
-				logrus.Errorf("Error when parse jid: %v", err)
-			} else {
-				pn, err := cli.Store.LIDs.GetPNForLID(ctx, lid)
-				if err != nil {
-					logrus.Errorf("Error when get pn for lid %s: %v", lid.String(), err)
-				}
-				if !pn.IsEmpty() {
-					if from_group != "" {
-						body["from"] = fmt.Sprintf("%s in %s", pn.String(), from_group)
-					} else {
-						body["from"] = pn.String()
-					}
-				}
-			}
-		}
-	}
-	if message.ID != "" {
-		tags := regexp.MustCompile(`\B@\w+`).FindAllString(message.Text, -1)
-		tagsMap := make(map[string]bool)
-		for _, tag := range tags {
-			tagsMap[tag] = true
-		}
-		for tag := range tagsMap {
-			lid, err := types.ParseJID(tag[1:] + "@lid")
-			if err != nil {
-				logrus.Errorf("Error when parse jid: %v", err)
-			} else {
-				pn, err := cli.Store.LIDs.GetPNForLID(ctx, lid)
-				if err != nil {
-					logrus.Errorf("Error when get pn for lid %s: %v", lid.String(), err)
-				}
-				if !pn.IsEmpty() {
-					message.Text = strings.Replace(message.Text, tag, fmt.Sprintf("@%s", pn.User), -1)
-				}
-			}
-		}
-		body["message"] = message
-	}
-	if pushname := evt.Info.PushName; pushname != "" {
-		body["pushname"] = pushname
-	}
-	if waReaction.Message != "" {
-		body["reaction"] = waReaction
-	}
-	if evt.IsViewOnce {
-		body["view_once"] = evt.IsViewOnce
-	}
-	if forwarded {
-		body["forwarded"] = forwarded
-	}
-	if timestamp := evt.Info.Timestamp.Format(time.RFC3339); timestamp != "" {
-		body["timestamp"] = timestamp
-	}
-
-	// Handle protocol messages (revoke, etc.)
-	if protocolMessage := evt.Message.GetProtocolMessage(); protocolMessage != nil {
-		protocolType := protocolMessage.GetType().String()
-
-		switch protocolType {
-		case "REVOKE":
-			body["action"] = "message_revoked"
-			if key := protocolMessage.GetKey(); key != nil {
-				body["revoked_message_id"] = key.GetID()
-				body["revoked_from_me"] = key.GetFromMe()
-				if key.GetRemoteJID() != "" {
-					body["revoked_chat"] = key.GetRemoteJID()
-				}
-			}
-		case "MESSAGE_EDIT":
-			body["action"] = "message_edited"
-			if editedMessage := protocolMessage.GetEditedMessage(); editedMessage != nil {
-				if editedText := editedMessage.GetExtendedTextMessage(); editedText != nil {
-					body["edited_text"] = editedText.GetText()
-				} else if editedConv := editedMessage.GetConversation(); editedConv != "" {
-					body["edited_text"] = editedConv
-				}
-			}
-		}
-	}
-
-	if audioMedia := evt.Message.GetAudioMessage(); audioMedia != nil {
-		path, err := utils.ExtractMedia(ctx, cli, config.PathMedia, audioMedia)
-		if err != nil {
-			logrus.Errorf("Failed to download audio from %s: %v", evt.Info.SourceString(), err)
-			return nil, pkgError.WebhookError(fmt.Sprintf("Failed to download audio: %v", err))
-		}
-		body["audio"] = path
-	}
-
-	if contactMessage := evt.Message.GetContactMessage(); contactMessage != nil {
-		body["contact"] = contactMessage
-	}
-
-	if documentMedia := evt.Message.GetDocumentMessage(); documentMedia != nil {
-		path, err := utils.ExtractMedia(ctx, cli, config.PathMedia, documentMedia)
-		if err != nil {
-			logrus.Errorf("Failed to download document from %s: %v", evt.Info.SourceString(), err)
-			return nil, pkgError.WebhookError(fmt.Sprintf("Failed to download document: %v", err))
-		}
-		body["document"] = path
-	}
-
-	if imageMedia := evt.Message.GetImageMessage(); imageMedia != nil {
-		path, err := utils.ExtractMedia(ctx, cli, config.PathMedia, imageMedia)
-		if err != nil {
-			logrus.Errorf("Failed to download image from %s: %v", evt.Info.SourceString(), err)
-			return nil, pkgError.WebhookError(fmt.Sprintf("Failed to download image: %v", err))
-		}
-		body["image"] = path
-	}
-
-	if listMessage := evt.Message.GetListMessage(); listMessage != nil {
-		body["list"] = listMessage
-	}
-
-	if liveLocationMessage := evt.Message.GetLiveLocationMessage(); liveLocationMessage != nil {
-		body["live_location"] = liveLocationMessage
-	}
-
-	if locationMessage := evt.Message.GetLocationMessage(); locationMessage != nil {
-		body["location"] = locationMessage
-	}
-
-	if orderMessage := evt.Message.GetOrderMessage(); orderMessage != nil {
-		body["order"] = orderMessage
-	}
-
-	if stickerMedia := evt.Message.GetStickerMessage(); stickerMedia != nil {
-		path, err := utils.ExtractMedia(ctx, cli, config.PathMedia, stickerMedia)
-		if err != nil {
-			logrus.Errorf("Failed to download sticker from %s: %v", evt.Info.SourceString(), err)
-			return nil, pkgError.WebhookError(fmt.Sprintf("Failed to download sticker: %v", err))
-		}
-		body["sticker"] = path
-	}
-
-	if videoMedia := evt.Message.GetVideoMessage(); videoMedia != nil {
-		path, err := utils.ExtractMedia(ctx, cli, config.PathMedia, videoMedia)
-		if err != nil {
-			logrus.Errorf("Failed to download video from %s: %v", evt.Info.SourceString(), err)
-			return nil, pkgError.WebhookError(fmt.Sprintf("Failed to download video: %v", err))
-		}
-		body["video"] = path
-	}
-
-	return body, nil
-}
-
-func submitWebhook(payload map[string]any, url string) error {
+func submitWebhook(_ context.Context, payload map[string]any, url string) error {
 	client := &http.Client{Timeout: 10 * time.Second}
 
 	postBody, err := json.Marshal(payload)
@@ -246,54 +51,4 @@ func submitWebhook(payload map[string]any, url string) error {
 	}
 
 	return pkgError.WebhookError(fmt.Sprintf("error when submit webhook after %d attempts: %v", attempt, err))
-}
-
-// forwardDeleteToWebhook sends a delete event to webhook
-func forwardDeleteToWebhook(ctx context.Context, evt *events.DeleteForMe, message *domainChatStorage.Message) error {
-	logrus.Info("Forwarding delete event to webhook:", config.WhatsappWebhook)
-	payload, err := createDeletePayload(ctx, evt, message)
-	if err != nil {
-		return err
-	}
-
-	for _, url := range config.WhatsappWebhook {
-		if err = submitWebhook(payload, url); err != nil {
-			return err
-		}
-	}
-
-	logrus.Info("Delete event forwarded to webhook")
-	return nil
-}
-
-// createDeletePayload creates a webhook payload for delete events
-func createDeletePayload(_ context.Context, evt *events.DeleteForMe, message *domainChatStorage.Message) (map[string]any, error) {
-	body := make(map[string]any)
-
-	// Basic delete event information
-	body["action"] = "message_deleted_for_me"
-	body["deleted_message_id"] = evt.MessageID
-	body["sender_id"] = evt.SenderJID.User
-	body["timestamp"] = time.Now().Format(time.RFC3339)
-
-	// Include original message information if available
-	if message != nil {
-		body["chat_id"] = message.ChatJID
-		body["original_content"] = message.Content
-		body["original_sender"] = message.Sender
-		body["original_timestamp"] = message.Timestamp.Format(time.RFC3339)
-		body["was_from_me"] = message.IsFromMe
-
-		if message.MediaType != "" {
-			body["original_media_type"] = message.MediaType
-			body["original_filename"] = message.Filename
-		}
-	}
-
-	// Parse sender JID for proper formatting
-	if evt.SenderJID.Server != "" {
-		body["from"] = evt.SenderJID.String()
-	}
-
-	return body, nil
 }
